requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pygame>=2.5.0
tensorboard>=2.14.0
scipy>=1.11.0
tqdm>=4.65.0
pyyaml>=6.0
shimmy>=1.3.0
